refactor, implementacja gcloud TTS, wyodrębniono tworznie pliku do osobnej funkcji

+++ speech_recognition.py
@@ -1,12 +1,14 @@
 from google.cloud.speech_v1 import types
-import io, os
+from google.cloud.speech_v1.services.speech import SpeechClient
+import io
+import os
 
 LANGUAGE_RECOGNITION = 'pl-PL'
 
 
-def speech_recognition(gcloud_client, hertz_rate):
+def speech_recognition(hertz_rate):
 
-    client = gcloud_client
+    client = SpeechClient()
 
     with io.open('file.wav', 'rb') as audio_file:
         content = audio_file.read()
@@ -21,7 +23,8 @@ def speech_recognition(gcloud_client, hertz_rate):
 
     response = client.recognize(config=config, audio=audio)
 
-    os.remove('file.wav')
+    if os.path.isfile('file.wav'):
+        os.remove('file.wav')
 
     for result in response.results:
         return result.alternatives[0].transcript
